Feat: 12단계 완료 모달 구현 및 Vote관련 모달 네이밍 수정
